Revised code to address issues raised by rubocop. Split measures_controller tests so composite tests are now in a separate file.
