+++ app/helpers/measure_helper.rb
@@ -46,67 +46,58 @@ module MeasureHelper
     end
   end
 
-  def self.update_measures(measures, params, current_user, measure_details, vsac_options, vsac_ticket_granting_ticket, is_update, existing)
+  def self.update_measures(measures, current_user, is_update, existing)
     measures.each do |measure|
-      if (!is_update)
+      if !is_update
         existing = CqlMeasure.by_user(current_user).where(hqmf_set_id: measure.hqmf_set_id).first
-        if !existing.nil?
-          measures.each {|m| m.delete}
+        unless existing.nil?
+          measures.each(&:delete)
           error_message = {title: "Error Loading Measure", summary: "A version of this measure is already loaded.", body: "You have a version of this measure loaded already.  Either update that measure with the update button, or delete that measure and re-upload it."}
           return error_message
         end
-      else
-        if measure.component
-          if existing.hqmf_set_id != measure.composite_hqmf_set_id
-            measures.each {|m| m.delete}
-            error_message = {title: "Error Updating Measure", summary: "The update file does not match the measure.", body: "You have attempted to update a measure with a file that represents a different measure.  Please update the correct measure or upload the file as a new measure."}
-            return error_message
-          end
-        else 
-          if existing.hqmf_set_id != measure.hqmf_set_id
-            measures.each {|m| m.delete}
-            error_message = {title: "Error Updating Measure", summary: "The update file does not match the measure.", body: "You have attempted to update a measure with a file that represents a different measure.  Please update the correct measure or upload the file as a new measure."}
-            return error_message
-          end
+      elsif measure.component
+        if existing.hqmf_set_id != measure.composite_hqmf_set_id
+          measures.each(&:delete)
+          error_message = {title: "Error Updating Measure", summary: "The update file does not match the measure.", body: "You have attempted to update a measure with a file that represents a different measure.  Please update the correct measure or upload the file as a new measure."}
+          return error_message
         end
+      elsif existing.hqmf_set_id != measure.hqmf_set_id
+        measures.each(&:delete)
+        error_message = {title: "Error Updating Measure", summary: "The update file does not match the measure.", body: "You have attempted to update a measure with a file that represents a different measure.  Please update the correct measure or upload the file as a new measure."}
+        return error_message
       end
 
       # Exclude patient birthdate and expired OIDs used by SimpleXML parser for AGE_AT handling and bad oid protection in missing VS check
       missing_value_sets = (measure.as_hqmf_model.all_code_set_oids - measure.value_set_oids - ['2.16.840.1.113883.3.117.1.7.1.70', '2.16.840.1.113883.3.117.1.7.1.309'])
-      if missing_value_sets.length > 0
-        measures.each {|m| m.delete}
-        error_message = {title: "Measure is missing value sets", summary: "The measure you have tried to load is missing value sets.", body: "The measure you are trying to load is missing value sets.  Try re-packaging and re-exporting the measure from the Measure Authoring Tool.  The following value sets are missing: [#{missing_value_sets.join(', ')}]"}
-        return error_message
-      end
+      next if missing_value_sets.empty?
+      measures.each(&:delete)
+      error_message = {title: "Measure is missing value sets", summary: "The measure you have tried to load is missing value sets.", body: "The measure you are trying to load is missing value sets.  Try re-packaging and re-exporting the measure from the Measure Authoring Tool.  The following value sets are missing: [#{missing_value_sets.join(', ')}]"}
+      return error_message
     end
-    if (existing && is_update)
+    if existing && is_update
       existing.component_hqmf_set_ids.each do |component_hqmf_set_id|
         component_measure = CqlMeasure.by_user(current_user).where(hqmf_set_id: component_hqmf_set_id).first
         component_measure.delete
       end
       existing.delete
     end
-    return nil
+    nil
   end
   
   def self.measures_population_update(measures, is_update, current_user, measure_details)
     measures.each do |measure|
       current_user.measures << measure
 
-      if (is_update)
+      if is_update
         measure.populations.each_with_index do |population, population_index|
-          population['title'] = measure_details['population_titles'][population_index] if (measure_details['population_titles'])
-        end
-        # check if episode ids have changed
-        if (measure.episode_of_care?)
-          keys = measure.data_criteria.values.map {|d| d['source_data_criteria'] if d['specific_occurrence']}.compact.uniq
+          population['title'] = measure_details['population_titles'][population_index] if measure_details['population_titles']
         end
       else
         measure.needs_finalize = measure.populations.size > 1
         if measure.populations.size > 1
           strat_index = 1
           measure.populations.each do |population|
-            if (population[HQMF::PopulationCriteria::STRAT])
+            if population[HQMF::PopulationCriteria::STRAT]
               population['title'] = "Stratification #{strat_index}"
               strat_index += 1
             end
@@ -128,5 +119,4 @@ module MeasureHelper
     end
     current_user.save!
   end
-
 end
